pyproject.toml
[tool.poetry]
name = "django-project-structure"
version = "0.1.0"
description = "django-project-structure"
authors = ["ArsenHrynevych <102324797+Arsen-hrynevych@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
